+++ tests/acceptance/suites/runtime-config.js
@@ -1,14 +1,15 @@
 import { PORT } from '../constants';
-import ParentRuntimeConfig from '../blocks/parentruntimeconfig';
-import ChildRuntimeConfig from '../blocks/childruntimeconfig';
+import RuntimeConfigWrapper from '../blocks/runtimeconfigwrapper';
 import Page from '../blocks/page';
 
 fixture`Runtime Config`
   .page(`http://localhost:${PORT}/iframe_test`)
 
 test('Parent runtime config is passed to the child', async t => {
-  await ParentRuntimeConfig.set('linkTarget', '_blank');
+  const parentRuntimeConfig = new RuntimeConfigWrapper({isParent: true});
+  const childRuntimeConfig = new RuntimeConfigWrapper({isParent: false});
+  await parentRuntimeConfig.set('linkTarget', '_blank');
   await Page.switchToIframe();
-  const linkTarget = await ChildRuntimeConfig.get('linkTarget');
+  const linkTarget = await childRuntimeConfig.get('linkTarget');
   await t.expect(linkTarget).eql('_blank');
 });
